# Reduce overhead of ReadOnlySequence<T>.CopyTo